+++ lib/spout/tests/domain_format.rb
@@ -2,7 +2,7 @@ module Spout
   module Tests
     module DomainFormat
 
-      def assert_domain_format(item, msg = nil)
+      def assert_domain_format(item)
         result = begin
           json = JSON.parse(File.read(item))
           if json.kind_of?(Array)
@@ -14,10 +14,9 @@ module Spout
           false
         end
 
-        full_message = build_message(msg, "Must be an array of choice hashes. Ex:\n[\n  {\n    \"value\":        \"1\",\n    \"display_name\": \"Option 1\",\n    \"description\":  \"...\"\n  },\n  { ... },\n  ...\n]")
-        assert_block(full_message) do
-          result
-        end
+        message = "Must be an array of choice hashes. Ex:\n[\n  {\n    \"value\":        \"1\",\n    \"display_name\": \"Option 1\",\n    \"description\":  \"...\"\n  },\n  { ... },\n  ...\n]"
+
+        assert result, message
       end
 
       Dir.glob("domains/**/*.json").each do |file|